fix: 默认的Amount类型改为decimal，并从ICardComponent中移除

+++ Component/CardComponent.cs
@@ -9,7 +9,7 @@ public abstract class CardComponent : ICardComponent
     public string ComponentId => CardComponentRegistry.GetComponentId(GetType());
 
     [ComponentState]
-    public virtual int Amount { get; set; }
+    public virtual decimal Amount { get; set; }
 
     public IComponentsCardModel? Card { get; private set; }
